+++ lib/Expression/Func/SumFunction.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace PhpBench\Expression\Func;
+
+use function array_sum;
+use PhpBench\Expression\Ast\ListNode;
+use PhpBench\Expression\Ast\PhpValue;
+use PhpBench\Expression\Ast\PhpValueFactory;
+
+final class SumFunction
+{
+    public function __invoke(ListNode $values): PhpValue
+    {
+        return PhpValueFactory::fromValue(array_sum($values->nonNullPhpValues()));
+    }
+}